lambda/poster/requirements.txt
python-dotenv==1.1.0
tweepy==4.15.0